refactor(sider): extract TeamSiderSection and reorganize Sider directory
- Extract team sidebar logic into TeamSiderSection component
- Move CronJobSiderItem + CronJobSiderSection into CronJobSiderSection/
- Move SiderToolbar, SiderSearchEntry, SiderScheduledEntry into SiderNav/
- Add barrel exports (index.ts) for both subdirectories
- Simplify Sider/index.tsx by delegating to extracted components
- Update test import paths and mocks to match new file locations
